Major Changes to the Order of Information
I rearranged a lot of the information that was originally included in this lesson. I focused on shorter bits of information followed by a short quiz at the beginning and then added in a longer section of quiz towards the end to incorporate all of the information learned. I also updated and rearranged the syntax to make it easier to read and understand when there is information being presented vs. when a student needs to make a choice/selection in the app.